Safari-609.1.11.1/LayoutTests/fast/media/getStartDate-NaN.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Tests that calling media.getStartDate() returns NaN when no start date is available.");

shouldBeTrue("isNaN(document.createElement('audio').getStartDate())");
shouldBeTrue("isNaN(document.createElement('video').getStartDate())");
shouldBeType("document.createElement('audio').getStartDate()", "Date");
shouldBeType("document.createElement('video').getStartDate()", "Date");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
