Safari-609.1.11.1/LayoutTests/fast/media/getStartDate-NaN-expected.txt

Tests that calling media.getStartDate() returns NaN when no start date is available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isNaN(document.createElement('audio').getStartDate()) is true
PASS isNaN(document.createElement('video').getStartDate()) is true
PASS document.createElement('audio').getStartDate() is an instance of Date
PASS document.createElement('video').getStartDate() is an instance of Date
PASS successfullyParsed is true

TEST COMPLETE
